% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\name{adapt_transparency}
\alias{adapt_transparency}
\title{Adapt transparency}
\usage{
adapt_transparency(ix, groups, colors, alpha_new = 0.25)
}
\arguments{
\item{ix}{index}

\item{groups}{list of cell types of interest}

\item{alpha_new}{value for new transparency (0 = transparent, 1 = full color)}
}
\description{
Adapt transparency
}
