% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{CELLTYPE_ANNOT_ABBREV_DICT}
\alias{CELLTYPE_ANNOT_ABBREV_DICT}
\title{Dictionary of cell type abbreviations}
\format{
An object of class \code{character} of length 14.
}
\usage{
CELLTYPE_ANNOT_ABBREV_DICT
}
\description{
Abbreviation for the different cell types to be used in visualization
}
\keyword{datasets}
