% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{CELLTYPES_COLOR_PALETTE}
\alias{CELLTYPES_COLOR_PALETTE}
\title{Dictionary of color palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
CELLTYPES_COLOR_PALETTE
}
\description{
Dictionary of color palettes
}
\keyword{datasets}
