% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{check_length_genesymbol}
\alias{check_length_genesymbol}
\title{Check genesymbol length}
\usage{
check_length_genesymbol(vec)
}
\arguments{
\item{vec}{vector with genesymbols}
}
\description{
Check genesymbol length
}
