% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{COLORS_L1}
\alias{COLORS_L1}
\title{Color palette for L1 cell types}
\format{
An object of class \code{character} of length 9.
}
\usage{
COLORS_L1
}
\description{
Color palette for L1 cell types
}
\keyword{datasets}
