% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{COLORS_L1_TME}
\alias{COLORS_L1_TME}
\title{Color palette for L1 TME cell types}
\format{
An object of class \code{character} of length 8.
}
\usage{
COLORS_L1_TME
}
\description{
Color palette for L1 TME cell types
}
\keyword{datasets}
