% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{COLORS_L1_TME_HIGHLIGHTED}
\alias{COLORS_L1_TME_HIGHLIGHTED}
\title{Color palette for L1 TME cell types (highlight groups of interest)}
\format{
An object of class \code{character} of length 8.
}
\usage{
COLORS_L1_TME_HIGHLIGHTED
}
\description{
Color palette for L1 TME cell types (highlight groups of interest)
}
\keyword{datasets}
