% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{COLORS_L2}
\alias{COLORS_L2}
\title{Color palette for L2 cell types}
\format{
An object of class \code{character} of length 10.
}
\usage{
COLORS_L2
}
\description{
Color palette for L2 cell types
}
\keyword{datasets}
