% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_bulk.R
\name{compute_corr}
\alias{compute_corr}
\title{Compute correlation}
\usage{
compute_corr(df, i, v1, v2, method = "pearson")
}
\arguments{
\item{df}{dataframe}

\item{i}{vector of indices}

\item{v1}{name of group1}

\item{v2}{name of group2}

\item{method}{method to use for computing correlation}
}
\description{
Compute correlation between two vectors in a dataframe
}
