% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{convert_iteration_matrix_to_df}
\alias{convert_iteration_matrix_to_df}
\title{Create a dataframe from a sampling grid}
\usage{
convert_iteration_matrix_to_df(i, sampling_grid)
}
\arguments{
\item{i}{The iteration number}

\item{sampling_grid}{A sampling grid}
}
\value{
A dataframe
}
\description{
Create a dataframe from a sampling grid
}
