% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_hm}
\alias{create_hm}
\title{Plot heatmap + save}
\usage{
create_hm(
  mat,
  col_fun = NULL,
  output_file = "heatmap.pdf",
  legend_title = "score",
  row_title = NULL,
  column_title = NULL,
  column_title_rot = 90,
  cluster_columns = TRUE,
  show_column_dend = TRUE,
  show_row_dend = TRUE,
  save_plot = FALSE,
  top_annotation = NULL,
  show_heatmap_legend = TRUE,
  heatmap_legend_param = list(),
  annotation_legend_side = "right",
  heatmap_legend_side = "right",
  merge_legend = TRUE,
  left_annotation = NULL,
  custom_cell_fun = NULL,
  custom_layer_fun = NULL,
  cluster_rows = TRUE,
  cell_size = 4,
  heatmap_legend_list = NULL,
  row_split = NULL,
  column_split = NULL
)
}
\arguments{
\item{mat}{matrix}

\item{col_fun}{color function}

\item{output_file}{output file name}

\item{legend_title}{legend title}

\item{column_title}{column title}

\item{column_title_rot}{rotation of column title}

\item{cluster_columns}{cluster columns}

\item{show_column_dend}{show column dendrogram}

\item{show_row_dend}{show row dendrogram}

\item{save_plot}{save plot}

\item{top_annotation}{column heatmap annotation}

\item{row_split}{null}

\item{column_split}{null}

\item{width}{width of the plot}

\item{height}{height of the plot}

\item{col}{color}
}
\value{
hm
}
\description{
Plot heatmap + save
}
