% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_lineplot}
\alias{create_lineplot}
\title{Create lineplot of average expression for ligand-receptor pairs}
\usage{
create_lineplot(interactions_df)
}
\arguments{
\item{interactions_df}{data.frame}
}
\value{
ggplot
}
\description{
Create lineplot of average expression for ligand-receptor pairs
}
