% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{custom_cell_function_triangular_default}
\alias{custom_cell_function_triangular_default}
\title{Cell function for ComplexHeatmap}
\usage{
custom_cell_function_triangular_default(j, i, x, y, width, height, fill)
}
\arguments{
\item{j}{column index}

\item{i}{row index}

\item{x}{x coordinate}

\item{y}{y coordinate}

\item{width}{width of the cell}

\item{height}{height of the cell}

\item{fill}{fill color}
}
\value{
NA
}
\description{
Cell function for ComplexHeatmap
}
