% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{db_filtering}
\alias{db_filtering}
\title{Filter and clean-up database}
\usage{
db_filtering(db)
}
\arguments{
\item{db}{dataframe with database of interactions, the following columns should at least be present 'source_genesymbol' and 'target_genesymbol'}
}
\description{
Filter and clean-up database
}
