% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering the Seurat object to reduce size of object
\enumerate{
\item Only keep genes that are part of the interactions db
\item Only keep cell types with at least N cells (user-defined)
}}
\usage{
filtering(seurat_obj, annot, min_cells, genes_oi)
}
\arguments{
\item{(Seurat)}{seurat_obj Seurat object}

\item{(character)}{annot Annotation to use for filtering}

\item{(integer)}{min_cells Minimum number of cells per annotation}

\item{(array)}{genes_oi list of genes of interest from database}
}
\value{
(Seurat) seurat_obj Filtered seurat object
}
\description{
Filtering the Seurat object to reduce size of object
\enumerate{
\item Only keep genes that are part of the interactions db
\item Only keep cell types with at least N cells (user-defined)
}
}
\examples{
seurat_obj <- filtering(seurat_obj, "custom_annot", 300, genes_oi)
}
