% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{format_cell2cell}
\alias{format_cell2cell}
\title{Format output of Cell2Cell}
\usage{
format_cell2cell(
  p,
  file_info_colnames = c("method", "sample_id", "region", "iteration")
)
}
\arguments{
\item{p}{path to file}
}
\value{
obj format
}
\description{
Format output of Cell2Cell
}
