% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{format_cellchat}
\alias{format_cellchat}
\title{Format output of CellChat}
\usage{
format_cellchat(
  p,
  file_info_colnames = c("method", "sample_id", "region", "iteration")
)
}
\arguments{
\item{p}{path to file}
}
\value{
obj formatted output
}
\description{
Format output of CellChat
}
