% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{format_liana}
\alias{format_liana}
\title{Format output of Liana}
\usage{
format_liana(
  p,
  file_info_colnames = c("method", "sample_id", "region", "iteration")
)
}
\arguments{
\item{p}{path to file}
}
\value{
obj formatted output
}
\description{
Format output of Liana
}
