% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_cell_dims}
\alias{get_cell_dims}
\title{Get size of heatmap (in mm)}
\usage{
get_cell_dims(matrix, cell_width = 0.1, cell_height = 0.1)
}
\arguments{
\item{matrix}{matrix}

\item{cell_width}{width of cell in mm, default=0.1 (square)}
}
\value{
list with height and width
}
\description{
use desired cell_width and cell_height to compute size of matrix
}
