% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_cell_function}
\alias{get_cell_function}
\title{Cell function}
\usage{
get_cell_function(matrix, is_upper_tri = FALSE, add_annot = TRUE)
}
\arguments{
\item{matrix}{matrix to use for annotation/coloring (can be different from actual matrix used for heatmap)}

\item{is_upper_tri}{heatmap only contains upper triangle of matrix (default=FALSE)}

\item{add_annot}{annotate cells in heatmap (default=TRUE)}
}
\description{
Cell function
}
