% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{get_file_info}
\alias{get_file_info}
\title{Get file information}
\usage{
get_file_info(p)
}
\arguments{
\item{p}{path to file}
}
\value{
file_info in array
}
\description{
Split the filename based on double underscores to extract several characteristics of the file.
}
\examples{
get_file_info("cell2cell__6245__Tumour_edge__0.csv")
}
