% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata for Cell2Cell}
\usage{
get_metadata(
  paths,
  file_info_colnames = c("method", "sample_id", "region", "iteration")
)
}
\arguments{
\item{paths}{paths to files}

\item{files_info_colnames}{column names for metadata}
}
\value{
metadata dataframe
}
\description{
Get metadata for Cell2Cell
}
