% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_name}
\alias{get_name}
\title{Extract name from filepath}
\usage{
get_name(filepath)
}
\arguments{
\item{filepath}{filepath}
}
\value{
name of file without extension
}
\description{
Extract name from filepath
}
\examples{
examples get_name("R/my_script.R") returns 'my_script'
}
