% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_optimal_output_size}
\alias{get_optimal_output_size}
\title{Obtain optimal size of heatmap or list of heatmaps}
\usage{
get_optimal_output_size(hm, m = 4)
}
\arguments{
\item{hm}{heatmap or list of heatmaps}

\item{m}{margin}
}
\value{
list of height and width
}
\description{
Obtain optimal size of heatmap or list of heatmaps
}
