% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc_ccis.R
\name{get_sample_ids}
\alias{get_sample_ids}
\title{Get sample IDs}
\usage{
get_sample_ids(metadata_df)
}
\arguments{
\item{metadata_df}{metadata dataframe}
}
\value{
sample_ids in array
}
\description{
Get sample IDs
}
