% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{hard_combine_dbs}
\alias{hard_combine_dbs}
\title{Hard combine databases}
\usage{
hard_combine_dbs(cellchat_db, cpdb_db, liana_db)
}
\arguments{
\item{cellchat_db}{cellchat database}

\item{cpdb_db}{cpdb_database}

\item{liana_db}{liana database}
}
\description{
Combine databases based on 'source_genesymbol' and 'target_genesymbol'
}
