% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_cellchat.R
\name{infer_cellchat}
\alias{infer_cellchat}
\title{Workflow for CellChat}
\usage{
infer_cellchat(seurat_obj, output_dir, args, i = NULL)
}
\arguments{
\item{(Seurat)}{seurat_obj Seurat object for a sample}

\item{(list)}{args List of user arguments}

\item{(integer)}{i Index of sampling grid}
}
\value{
(list) List of CellChat results
}
\description{
Workflow for CellChat
}
\examples{
infer_cellchat(seurat_obj, args) or infer_interactions(seurat_obj, args, i = 1)
}
