% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{init_logging}
\alias{init_logging}
\title{Set up logging}
\usage{
init_logging(log_level, log_file = NULL)
}
\arguments{
\item{log_level}{level of logging (1-5)}

\item{log_file}{path to log file (optional)}
}
\value{
logger object
}
\description{
Set up logging
}
\examples{
logr <- init_logging(3)
}
