% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{INTERACTIONS_POST_FILTERING_OPTIONS}
\alias{INTERACTIONS_POST_FILTERING_OPTIONS}
\title{List of filtering options}
\format{
An object of class \code{character} of length 2.
}
\usage{
INTERACTIONS_POST_FILTERING_OPTIONS
}
\description{
Types of filtering
Count the number of interactions between cell type groups
}
\details{
CONSTANT VARIABLES USED IN MULTIPLE SCRIPTS
}
\keyword{datasets}
