% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{L1_TME}
\alias{L1_TME}
\title{L1 TME cell type labels (TME only, no tumor)}
\format{
An object of class \code{character} of length 8.
}
\usage{
L1_TME
}
\description{
L1 TME cell type labels (TME only, no tumor)
}
\keyword{datasets}
