% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_color_palette.R
\docType{data}
\name{L2_TME}
\alias{L2_TME}
\title{L2 TME cell type labels (TME only, no tumor)}
\format{
An object of class \code{character} of length 8.
}
\usage{
L2_TME
}
\description{
L2 TME cell type labels (TME only, no tumor)
}
\keyword{datasets}
