% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{load_cellchat_db}
\alias{load_cellchat_db}
\title{Load CellChat DB}
\usage{
load_cellchat_db()
}
\value{
dataframe with 'source_genesymbol' and 'target_genesymbol'
}
\description{
load cellcat database and format for merging
}
