% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_debug}
\alias{log_debug}
\title{Logging functions: log_debug}
\usage{
log_debug(...)
}
\arguments{
\item{...}{message}

\item{logr}{logger object}
}
\description{
Logging functions: log_debug
}
\examples{
log_debug("Hello world!")

}
