% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_error}
\alias{log_error}
\title{Logging functions: log_error}
\usage{
log_error(...)
}
\arguments{
\item{...}{message}

\item{logr}{logger object}
}
\description{
Logging functions: log_error
}
\examples{
log_error("Hello world!")

}
