% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_fatal}
\alias{log_fatal}
\title{Logging functions: log_fatal}
\usage{
log_fatal(...)
}
\arguments{
\item{...}{message}

\item{logr}{logger object}
}
\description{
Logging functions: log_fatal
}
\examples{
log_fatal("Hello world!")

}
