% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_info}
\alias{log_info}
\title{Logging functions: log_info}
\usage{
log_info(...)
}
\arguments{
\item{...}{message}

\item{logr}{logger object}
}
\description{
Logging functions: log_info
}
\examples{
log_info("Hello world!")
}
