% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_warn}
\alias{log_warn}
\title{Logging functions: log.warn}
\usage{
log_warn(...)
}
\arguments{
\item{...}{message}

\item{logr}{logging object}
}
\description{
Logging functions: log.warn
}
\examples{
log_warn("Hello world!")
}
