% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{lookup_uniprot}
\alias{lookup_uniprot}
\title{Look up uniprot based on genesymbol}
\usage{
lookup_uniprot(gene, gene_uniprot_table)
}
\arguments{
\item{gene}{genesymbol}

\item{gene_uniprot_table}{dataframe containing at least 'genesymbol' and 'uniprot' columns}
}
\description{
Look up uniprot based on genesymbol
}
