% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{make_complex}
\alias{make_complex}
\title{make complex}
\usage{
make_complex(vec, collapse = ":")
}
\arguments{
\item{vec}{vector}
}
\description{
Merge items in vector by collapse character specified by user, default = ':'
}
