% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{order_vector}
\alias{order_vector}
\title{order vector}
\usage{
order_vector(vec)
}
\arguments{
\item{vec}{vector}
}
\description{
order elements in vector alphabetically and return vector with same length
}
