% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pipeline_heatmap}
\alias{pipeline_heatmap}
\title{Pipeline for creating heatmaps (customized for 401_heatmap)}
\usage{
pipeline_heatmap(
  interactions_mvoted,
  pairs,
  obj_literature,
  setname = "TAM-BDM - Malignant",
  idx = 2
)
}
\arguments{
\item{interactions_mvoted}{interactions}

\item{pairs}{pairs of interest}

\item{setname}{name of the set}

\item{idx}{index of the column to use for the title}
}
\value{
list of heatmaps
}
\description{
Pipeline for creating heatmaps (customized for 401_heatmap)
}
