% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_expr_hist_ccis}
\alias{plot_expr_hist_ccis}
\title{Create lineplot of average expression for ligand-receptor pairs}
\usage{
plot_expr_hist_ccis(avg_expr, annot, args)
}
\arguments{
\item{avg_expr}{data.frame}

\item{annot}{data.frame}

\item{args}{list of arguments}
}
\value{
list of ggplots
}
\description{
Create lineplot of average expression for ligand-receptor pairs
}
