% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{rank_interactions}
\alias{rank_interactions}
\title{Ranking interactions for consensus using all tools}
\usage{
rank_interactions(
  cell_type_pair,
  connectome,
  sca,
  cytotalk,
  natmi,
  logfc,
  obj_cell2cell,
  obj_cellchat,
  obj_cpdb
)
}
\arguments{
\item{cell_type_pair}{cell type of interest for filtering}

\item{connectome}{dataframe with connectome results (interactions ordered by descending score)}

\item{sca}{dataframe with sca results (interactions ordered by descending score)}

\item{cytotalk}{dataframe with cytotalk results (interactions ordered by descending score)}

\item{natmi}{dataframe with natmi results (interactions ordered by descending score)}

\item{logfc}{dataframe with logfc results (interactions ordered by descending score)}

\item{obj_cell2cell}{dataframe with cell2cell results (interactions ordered by (ascending) p-values)}

\item{obj_cellchat}{dataframe with cellchat results (interactions ordered by (ascending) p-values)}

\item{obj_cpdb}{dataframe with cellphonedb results (interactions ordered by (ascending) p-values)}
}
\description{
Ranking interactions for consensus using all tools
}
