% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\docType{data}
\name{REGION_GROUPED_LEVELS}
\alias{REGION_GROUPED_LEVELS}
\title{List of regions in GBM data}
\format{
An object of class \code{character} of length 3.
}
\usage{
REGION_GROUPED_LEVELS
}
\description{
Levels for the Region_grouped variable (ordered)
}
\keyword{datasets}
