% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_bulk.R
\name{run_bootstrap_corr}
\alias{run_bootstrap_corr}
\title{Run bootstrap correlation analysis}
\usage{
run_bootstrap_corr(
  pair,
  df,
  stat_func = compute_corr,
  n_iter = 500,
  method = "pearson"
)
}
\arguments{
\item{pair}{named vector of length 2 (names: Var1 and Var2) with columns to select for analysis}

\item{df}{dataframe}

\item{stat_func}{function to use for computation of statistic, i.e. correlation}

\item{n_iter}{number of iterations to use for bootstrapping}

\item{method}{method to use for computing correlation}
}
\description{
Run bootstrap correlation analysis
}
