% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_cells}
\alias{sample_cells}
\title{Sample cells from a Seurat object for a given annotation}
\usage{
sample_cells(annot, annot_name, n_cells, seurat_obj)
}
\arguments{
\item{annot}{An annotation to sample cells from (label, e.g. "Tumor")}

\item{annot_name}{The name of the annotation column in the Seurat object (e.g. "custom_annot")}

\item{n_cells}{The number of cells to sample for each annotation}

\item{seurat_obj}{A Seurat object}
}
\value{
A vector of cell IDs
}
\description{
Sample cells from a Seurat object for a given annotation
}
