% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_iteration}
\alias{sample_iteration}
\title{Sample cells for an iteration (or run)}
\usage{
sample_iteration(i, annot_name, cell_types, n_cells, seurat_obj)
}
\arguments{
\item{i}{The iteration number}

\item{annot_name}{The name of the annotation column in the Seurat object (e.g. "custom_annot")}

\item{cell_types}{A vector of cell types to sample from}

\item{n_cells}{The number of cells to sample for each annotation}

\item{seurat_obj}{A Seurat object}
}
\value{
a square matrix (n_cells x n_cells), columns = cell types, rows = cell IDs
}
\description{
Sample cells for an iteration (or run)
}
