% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_hm}
\alias{save_hm}
\title{Save heatmap}
\usage{
save_hm(
  hm_obj,
  annotation_legend_side = "right",
  heatmap_legend_side = "right",
  merge_legend = TRUE,
  heatmap_legend_list = NULL,
  annotation_legend_list = NULL,
  output_file = "heatmap.pdf"
)
}
\arguments{
\item{hm_obj}{heatmap created with ComplexHeatmap or create_hm()}

\item{annotation_legend_side}{position of annotation legend by default = 'right'}

\item{heatmap_legend_side}{position of annotation legend by default = 'right'}

\item{merge_legend}{merge legends, by default = TRUE}

\item{heatmap_legend_list}{list of custom legends, by default = NULL}

\item{annotation_legend_list}{list of custom legends, by default = NULL}

\item{output_file}{path for output file by default = "heatmap.pdf"}
}
\description{
Save heatmap with optimal output size (removing redundant whitespace)
}
