% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{setup_default_argparser}
\alias{setup_default_argparser}
\title{Setup default argparser}
\usage{
setup_default_argparser(description = "", default_output = "output")
}
\arguments{
\item{description}{Description for script}

\item{output_default}{Default output directory}
}
\value{
parser object
}
\description{
Set up a default argument parser, with two default arguments: log_level and output_dir
}
\examples{
setup_default_argparser(description = "Example")
}
