% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{update_cellchat_db}
\alias{update_cellchat_db}
\title{Update CellChat DB using (updated) CellPhoneDB}
\usage{
update_cellchat_db(cpdb_dir, db, output_dir)
}
\arguments{
\item{cpdb_dir}{directory with CellPhoneDB files}

\item{db}{database}

\item{output_dir}{path to directory for saving the generated CellPhoneDB database (.rds)}
}
\description{
Update CellChat DB using (updated) CellPhoneDB
}
