% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db_utils.R
\name{update_cpdb_db}
\alias{update_cpdb_db}
\title{Update CellPhoneDB}
\usage{
update_cpdb_db(
  source_cpdb_dir,
  db,
  gene_info_table,
  output_dir,
  return_list = FALSE
)
}
\arguments{
\item{source_cpdb_dir}{directory with CellPhoneDB files}

\item{db}{database with at least 'source_genesymbol' and 'target_genesymbol' and their subunits starting with 'source_genesymbol_subunit' or 'target_genesymbol_subunit'}

\item{gene_info_table}{dataframe with uniprot, protein name, genesymbol and ensembl}

\item{output_dir}{path to directory for saving the generated CellPhoneDB files}

\item{return_list}{return list with generated dataframes (default = FALSE)}
}
\value{
if return_list = TRUE, return list with complex_input, gene_input, interaction_input and protein_input
}
\description{
Update CellPhoneDB and generate required files
}
