% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_hm}
\alias{create_hm}
\title{Plot heatmap + save}
\usage{
create_hm(matrix, cell_width = 4, cell_height = 4, ...)
}
\arguments{
\item{matrix}{A matrix. Either numeric or character. If it is a simple vector, it will be converted to a one-column matrix.}

\item{mat}{matrix}

\item{col_fun}{color function}

\item{output_file}{output file name}

\item{legend_title}{legend title}

\item{save_plot}{save plot}
}
\value{
hm
}
\description{
Plot heatmap + save
}
