% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{filtering}
\alias{filtering}
\title{Filtering the Seurat object to reduce size of object}
\usage{
filtering(seurat_obj, annot, min_cells)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{annot}{Annotation to use for filtering}

\item{min_cells}{Minimum number of cells per annotation}

\item{genes_oi}{list of genes of interest from database}
}
\value{
seurat_obj Filtered seurat object
}
\description{
\enumerate{
\item Only keep genes that are part of the interactions db, 2. Only keep cell types with at least N cells (user-defined)
}
}
\examples{
dontrun{seurat_obj <- filtering(seurat_obj, "custom_annot", 300, genes_oi)}
}
