% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_cell_function}
\alias{get_cell_function}
\title{Obtain optimal ComplexHeatmap Size}
\usage{
get_cell_function(matrix, is_upper_tri = FALSE, add_annot = TRUE)
}
\arguments{
\item{matrix}{matrix to use}

\item{is_upper_tri}{default=FALSE,}

\item{add_annot}{default= TRUE}
}
\description{
Obtain optimal ComplexHeatmap Size
}
